--- .nyc_output/processinfo/3cad5cfc-06b4-4f72-af9b-8bf54be81b92.json
@@ -1,1 +1,0 @@
-{"uuid":"3cad5cfc-06b4-4f72-af9b-8bf54be81b92","parent":"47a50539-cba9-4545-a985-7ef71008628b","pid":17364,"argv":["C:\\Program Files\\nodejs\\node.exe","D:\\Property-Pro-Lite\\node_modules\\mocha\\bin\\_mocha","--timeout","0","--exit","--no-config","--no-package","--no-opts","--diff","--extension","js","--reporter","spec","--slow","75","--ui","bdd"],"execArgv":[],"cwd":"D:\\Property-Pro-Lite","time":1562861235161,"ppid":16548,"root":"3b6ded1b-507b-409f-b722-e7abf835d9d5","coverageFilename":"D:\\Property-Pro-Lite\\.nyc_output\\3cad5cfc-06b4-4f72-af9b-8bf54be81b92.json","files":["D:\\Property-Pro-Lite\\server\\index.js","D:\\Property-Pro-Lite\\server\\routes\\users.js","D:\\Property-Pro-Lite\\server\\controllers\\users.js","D:\\Property-Pro-Lite\\server\\models\\User.js","D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\userRegistration.js","D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\is-empty.js","D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\login.js","D:\\Property-Pro-Lite\\server\\helpers\\responses.js","D:\\Property-Pro-Lite\\server\\routes\\properties.js","D:\\Property-Pro-Lite\\server\\controllers\\properties.js","D:\\Property-Pro-Lite\\server\\models\\Property.js","D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\properties.js"]}
