--- .nyc_output/processinfo/47a50539-cba9-4545-a985-7ef71008628b.json
@@ -1,1 +1,0 @@
-{"uuid":"47a50539-cba9-4545-a985-7ef71008628b","parent":null,"pid":16548,"argv":["C:\\Program Files\\nodejs\\node.exe","D:\\Property-Pro-Lite\\node_modules\\mocha\\bin\\mocha","--no-timeout","--exit"],"execArgv":[],"cwd":"D:\\Property-Pro-Lite","time":1562861230959,"ppid":15820,"root":"3b6ded1b-507b-409f-b722-e7abf835d9d5","coverageFilename":"D:\\Property-Pro-Lite\\.nyc_output\\47a50539-cba9-4545-a985-7ef71008628b.json","files":[]}
