+++ .nyc_output/processinfo/6b4fbd13-984e-45ac-a42e-7a10c2b65ae7.json
@@ -1,0 +1,1 @@
+{"uuid":"6b4fbd13-984e-45ac-a42e-7a10c2b65ae7","parent":"f60285e4-40ff-4296-b521-d9f790a06bbd","pid":21884,"argv":["C:\\Program Files\\nodejs\\node.exe","D:\\Property-Pro-Lite\\node_modules\\mocha\\bin\\_mocha","--timeout","0","--exit","--no-config","--no-package","--no-opts","--diff","--extension","js","--reporter","spec","--slow","75","--ui","bdd"],"execArgv":[],"cwd":"D:\\Property-Pro-Lite","time":1562959394897,"ppid":14032,"root":"cf5410a4-903c-4ac6-84b6-fdb5bbf46e8e","coverageFilename":"D:\\Property-Pro-Lite\\.nyc_output\\6b4fbd13-984e-45ac-a42e-7a10c2b65ae7.json","files":["D:\\Property-Pro-Lite\\server\\index.js","D:\\Property-Pro-Lite\\server\\routes\\users.js","D:\\Property-Pro-Lite\\server\\controllers\\users.js","D:\\Property-Pro-Lite\\server\\models\\User.js","D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\userRegistration.js","D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\is-empty.js","D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\login.js","D:\\Property-Pro-Lite\\server\\helpers\\responses.js","D:\\Property-Pro-Lite\\server\\routes\\properties.js","D:\\Property-Pro-Lite\\server\\controllers\\properties.js","D:\\Property-Pro-Lite\\server\\models\\Property.js","D:\\Property-Pro-Lite\\server\\models\\propertiesType.js","D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\properties.js","D:\\Property-Pro-Lite\\server\\helpers\\errors.js"]}
