+++ .nyc_output/processinfo/f60285e4-40ff-4296-b521-d9f790a06bbd.json
@@ -1,0 +1,1 @@
+{"uuid":"f60285e4-40ff-4296-b521-d9f790a06bbd","parent":null,"pid":14032,"argv":["C:\\Program Files\\nodejs\\node.exe","D:\\Property-Pro-Lite\\node_modules\\mocha\\bin\\mocha","--no-timeout","--exit"],"execArgv":[],"cwd":"D:\\Property-Pro-Lite","time":1562959386259,"ppid":22332,"root":"cf5410a4-903c-4ac6-84b6-fdb5bbf46e8e","coverageFilename":"D:\\Property-Pro-Lite\\.nyc_output\\f60285e4-40ff-4296-b521-d9f790a06bbd.json","files":[]}
