+++ .nyc_output/processinfo/index.json
@@ -1,1 +1,1 @@
-{"processes":{"3cad5cfc-06b4-4f72-af9b-8bf54be81b92":{"parent":"47a50539-cba9-4545-a985-7ef71008628b","children":[]},"47a50539-cba9-4545-a985-7ef71008628b":{"parent":null,"children":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"]}},"files":{"D:\\Property-Pro-Lite\\server\\index.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\routes\\users.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\controllers\\users.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\models\\User.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\userRegistration.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\is-empty.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\login.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\helpers\\responses.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\routes\\properties.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\controllers\\properties.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\models\\Property.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"],"D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\properties.js":["3cad5cfc-06b4-4f72-af9b-8bf54be81b92"]},"externalIds":{}}
+{"processes":{"6b4fbd13-984e-45ac-a42e-7a10c2b65ae7":{"parent":"f60285e4-40ff-4296-b521-d9f790a06bbd","children":[]},"f60285e4-40ff-4296-b521-d9f790a06bbd":{"parent":null,"children":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"]}},"files":{"D:\\Property-Pro-Lite\\server\\index.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\routes\\users.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\controllers\\users.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\models\\User.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\userRegistration.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\is-empty.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\login.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\helpers\\responses.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\routes\\properties.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\controllers\\properties.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\models\\Property.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\models\\propertiesType.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\MIDDLEWARE\\properties.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"],"D:\\Property-Pro-Lite\\server\\helpers\\errors.js":["6b4fbd13-984e-45ac-a42e-7a10c2b65ae7"]},"externalIds":{}}
