+++ README.md
@@ -1,7 +1,7 @@
 # Property-Pro-Lite
 [![Build Status](https://travis-ci.com/Jordybastien/Property-Pro-Lite.svg?branch=develop)](https://travis-ci.com/Jordybastien/Property-Pro-Lite)
 
-[![Coverage Status](https://coveralls.io/repos/github/Jordybastien/Property-Pro-Lite/badge.svg?branch=develop)](https://coveralls.io/github/Jordybastien/Property-Pro-Lite?branch=develop)
+[![Coverage Status](https://coveralls.io/repos/github/Jordybastien/Property-Pro-Lite/badge.svg?branch=Travis)](https://coveralls.io/github/Jordybastien/Property-Pro-Lite?branch=Travis)
 
 [![Maintainability](https://api.codeclimate.com/v1/badges/5a331f0fe6f364594f9d/maintainability)](https://codeclimate.com/github/Jordybastien/Property-Pro-Lite/maintainability)
